+++ lib/constants.dart
@@ -1,4 +1,1 @@
 import 'package:flutter/material.dart';
-
-const TextStyle uiTextStyle = TextStyle(fontSize: 24);
-const TextStyle mediumTextStyle = TextStyle(fontSize: 20);