+++ lib/models/song_info.dart
@@ -1,0 +1,6 @@
+class SongInfo {
+  final String albumName;
+  final String songName;
+
+  SongInfo({required this.albumName, required this.songName});
+}